Boot the engine on Linux without FMOD
BitBuster links and runs: POSIX file system, Vulkan on the discrete GPU
with a 3-image swapchain, silent audio, and a stable main loop.

Audio: NullAudioContext and a silent FMODSoundReference stand in when
VOXAGINE_ENABLE_FMOD is off, so the engine runs without the proprietary
SDK. ResourceManager names FMODSoundReference concretely, so the stub
keeps the class and replaces only its behaviour.

Input: DirectXTK's Keyboard and Mouse join GamePad in being replaced by
SDL3-backed versions that keep DirectXTK's State field names. That
matters most for the keyboard, where 109 key mappings port untouched
instead of being retyped. Wheel movement is scaled back to DirectXTK's
120-per-notch units so tuning still holds.

Two real crashes fixed, both pre-existing:
- Application dereferenced the main camera unconditionally, three lines
  after carefully null-checking it into existence. Twice - the block is
  duplicated across the fixed and variable step paths, and fixing only
  the first still crashed in the second.
- ImGui's font atlas was never built, because that lived in the DX12
  context. Built lazily on the first frame, since Platform constructs
  the ImContext before ImguiSystem calls ImGui::CreateContext.

Also: my GameSources glob was compiling four files Game.vcxproj
excluded. SpawnerEntity calls OnDrawGizmos on a Component, which only
Entity declares, so it can never have built. Matched to the original
set rather than repairing dead code.

Removed three files carrying Sony's "SIE CONFIDENTIAL" PS4 SDK header -
user_new.cpp, user_malloc.cpp and ORBLinearAllocator.h. All were dead
with ORBIS gone. PlatformData.h was almost entirely _ORBIS and is now
the one field the engine reads.

Backslash include paths, TRUE/FALSE, and a wrong relative include for
Monster.h also cleared.

+++ Game/Source/Humanoids/Enemies/UmbrellaMonster.cpp
@@ -99,7 +99,7 @@ void UmbrellaMonster::MeleeAttack(Vector3& velocity)
 		float height = GetTransform()->GetPosition().y - m_pClosestTarget->GetTransform()->GetPosition().y;
 		height -= 10.f;
 
-		float initialVelocity = (1.f / cos(0.7f)) * sqrt(std::max((0.5f * glm::length(PhysicsBody::GRAVITY) * distance * distance) / (distance * tan(0.7f) + height), 1.f));
+		float initialVelocity = (1.f / cos(0.7f)) * sqrt(std::max(static_cast<float>((0.5f * glm::length(PhysicsBody::GRAVITY) * distance * distance) / (distance * tan(0.7f) + height)), 1.f));
 		Vector3 desiredVelocity = Vector3(0, initialVelocity * sin(0.7f) * m_fJumpForce.y, initialVelocity * cos(0.7f) * m_fJumpForce.x / m_fJumpForce.y);
 		desiredVelocity.y = std::min(desiredVelocity.y, m_fMaxJumpVelocity.y);
 		desiredVelocity.z = std::min(std::abs(desiredVelocity.z), m_fMaxJumpVelocity.x) * (desiredVelocity.z / std::abs(desiredVelocity.z));
